projects/qrcode_generator/requirements.txt
pyzbar
qrtools
pyqrcode
Pillow
pypng